revise parallel API so that it suits reduce() better and is simpler

+++ js/src/vm/threadpool.h
@@ -28,12 +28,12 @@ namespace js {
 
 class ThreadPoolWorker;
 
-typedef void (*TaskFun)(void *userdata, size_t workerId, uintptr_t stackLimit);
+typedef void (*TaskFun)(void *userdata, uint32_t workerId, uintptr_t stackLimit);
 
 class TaskExecutor
 {
 public:
-    virtual void executeFromWorker(size_t workerId, uintptr_t stackLimit) = 0;
+    virtual void executeFromWorker(uint32_t workerId, uintptr_t stackLimit) = 0;
 };
 
 /*
@@ -76,7 +76,7 @@ private:
     js::Vector<ThreadPoolWorker*, 8, SystemAllocPolicy> workers_;
 
     // Next worker for |submitOne()|. Atomically modified.
-    size_t nextId_;
+    uint32_t nextId_;
 
     void terminateWorkers();
 
@@ -87,7 +87,7 @@ public:
     bool init();
 
     // Return number of worker threads in the pool.
-    size_t numWorkers() { return workers_.length(); }
+    uint32_t numWorkers() { return workers_.length(); }
 
     // See comment on class:
     bool submitOne(TaskExecutor *executor);
